+++ client/ayon_core/tools/utils/__init__.py
@@ -37,6 +37,7 @@ from .lib import (
     get_qt_app,
     get_ayon_qt_app,
     get_openpype_qt_app,
+    get_qt_icon,
 )
 
 from .models import (
@@ -96,6 +97,7 @@ __all__ = (
     "get_qt_app",
     "get_ayon_qt_app",
     "get_openpype_qt_app",
+    "get_qt_icon",
 
     "RecursiveSortFilterProxyModel",
